+++ src/main/java/cart/dao/entity/CartEntity.java
@@ -2,33 +2,33 @@ package cart.dao.entity;
 
 public class CartEntity {
 
-    private final int productId;
-    private final int memberId;
+    private final Long productId;
+    private final Long memberId;
 
-    private CartEntity(final int memberId, final int productId) {
+    private CartEntity(final Long memberId, final Long productId) {
         this.memberId = memberId;
         this.productId = productId;
     }
 
-    public int getProductId() {
+    public Long getProductId() {
         return productId;
     }
 
-    public int getMemberId() {
+    public Long getMemberId() {
         return memberId;
     }
 
     public static class Builder {
 
-        private int productId;
-        private int memberId;
+        private Long productId;
+        private Long memberId;
 
-        public Builder productId(final int productId) {
+        public Builder productId(final Long productId) {
             this.productId = productId;
             return this;
         }
 
-        public Builder memberId(final int memberId) {
+        public Builder memberId(final Long memberId) {
             this.memberId = memberId;
             return this;
         }
